run clients as node modules, instead of processes
if client is implemented in node, we do not need
to spawn a separate client process but instead
require the node module directly from within
i3bar.js. thus saving system resources, because
we are running fewer processes.

(but this may be more fragile, since a fault in
a client may affect the whole bar process)